requirements.txt
mysql-connector-python==0.3.2-devel
nose==1.1.2
numpy==1.10.1
pandas==0.17.0
rpy2==2.7.4
scikit-fuzzy==0.1.3
scikit-learn==0.17
scikits.statsmodels==0.3.1
scipy==0.9.0
